refactor: Move MemTable to mod.rs following idiomatic Rust patterns
- Move MemTable implementation from mem_table.rs to mod.rs
- Keep skip_list.rs as supporting implementation
- Follow same pattern as sstable module structure
- All tests passing, maintains public API compatibility
